ISSUE_243

+++ isup/isup-api/src/main/java/org/mobicents/protocols/ss7/isup/message/parameter/FacilityIndicator.java
@@ -1,6 +1,6 @@
 /*
- * JBoss, Home of Professional Open Source
- * Copyright 2011, Red Hat, Inc. and individual contributors
+ * TeleStax, Open Source Cloud Communications
+ * Copyright 2012, Telestax Inc and individual contributors
  * by the @authors tag. See the copyright.txt in the distribution for a
  * full listing of individual contributors.
  *
@@ -40,9 +40,9 @@ public interface FacilityIndicator extends ISUPParameter {
     /**
      * See Q.763 3.22 facility indicator user-to-user service
      */
-    int _FACILITY_INDICATOR_UTUS = 2;
+    byte _FACILITY_INDICATOR_UTUS = 2;
 
-    int getFacilityIndicator();
+    byte getFacilityIndicator();
 
-    void setFacilityIndicator(int facilityIndicator);
+    void setFacilityIndicator(byte facilityIndicator);
 }
